backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend for project nsql1h26-anim"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.12.0",
    "fastapi>=0.135.3",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.43.0",
]